Install Rubocop
https://trello.com/c/otSqDTOR/2-install-rubocop

When you run the `Rubocop` command it combs through the whole codebase and
picks up on anything that breaks any of its current rules. Sometimes there are
exceptions and these should be added to the ignore list, or we can ignore
whole rules entirely if we disagree with them or feel they don't apply to us.

I've set this up to ignore some files that are Rails default or third party as
we shouldn't really need to modify those.