Install dotenv
https://trello.com/c/ZMUKLcZU/3-install-dotenv

Dotenv allows you to create a file called .env where you can store environment
variables.

This file is in .gitignore so it won't be committed to the repository or
publicly visible. This is for storing private data like passwords and private
keys for accessing APIs and it's so that people viewing our public repository
can't gain control of our accounts.

When setting up as a new developer on the app, you'll need to manually create
a .env file in the root directory and add the keys you need to this, someone
will need to help you with this. If you have access to Heroku you can copy
them from there.

Environment variables are accessed like this: `ENV['MEETUP_API_KEY']` for a
key saved in .env as `MEETUP_API_KEY=123`.