Install Pry
https://trello.com/c/UoC7iThK/4-install-pry

I prefer Pry for debugging over the Rails default 'Byebug' as it's what I'm
used to.

You can add `binding.pry` to any line in the codebase and when Rails hits that
line either in initialising the server or when you visit a page on the app, it
will freeze. When you check the server you'll see where it's paused at and can
run commands to see what the state of play is at that point in time.

This is useful for checking what is saved as variables (if anything) and what
data they have on them so far and can often help you realise why something
isn't working as it should, or to know what data is available to you at that
time.