Create Flexbox mixin
This keeps our flex code clean, allowing you to @include flex for a standard
flex container, but can pass variables to adjust attributes as you need to.

I've found I tend to only need to change one or two attributes about it
whenever I've used this so it's a bit easier to read CSS classes as this
removes all the flex clutter.