Add content_for additional stylesheet rendered
This allows you to add custom CSS to be rendered inline as the page loads,
which will take the highest priority. This allows you to use Ruby variables in
your CSS for doing things like adding a custom background image based on the
markup on the page without having to duplicate CSS in the view.